+++ src/main/java/de/maxhenkel/car/integration/jei/PainterRecipeCategoryYellow.java
@@ -1,5 +1,5 @@
 package de.maxhenkel.car.integration.jei;
-/*
+
 import de.maxhenkel.car.Main;
 import de.maxhenkel.car.items.ModItems;
 import mezz.jei.api.constants.VanillaTypes;
@@ -65,4 +65,3 @@ public class PainterRecipeCategoryYellow implements IRecipeCategory<PainterRecip
 
     }
 }
-*/